yolo5/requirements.txt
flask
pyyaml
loguru
pymongo
boto3
pathlib
detect
uuid
loguru

# testing

pylint
pytest